requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
qdrant-client==0.9.2
fastembed==0.3.4
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.2
fastapi-cors-middleware==0.1.0
python-jose==3.3.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
ujson==5.7.0
loguru==0.7.0
phonenumbers==8.13.34
sqlalchemy==1.4.46
alembic==1.13.1
psycopg2-binary==2.9.6
passlib[bcrypt]==1.7.4
